Fix issues reported by github

+++ jxrtestlib/JXRTestBmp.c
@@ -211,7 +211,7 @@ ERR PKImageEncode_WritePixels_BMP(
 
     for (i = cLine - 1; 0 <= i; --i)
     {
-        size_t offM = cbStride * i;
+        size_t offM = (size_t)cbStride * i;
         size_t offS = (size_t)cbLineS * (pIE->uHeight - (pIE->idxCurrentLine + i + 1));
 
         Call(pS->SetPos(pS, pIE->offPixel + offS));
